// client/src/state/actions-types/computer.action-types.ts
export enum ComputerActionsTypes {
    FETCH_COMPUTERS_REQUEST = 'FETCH_COMPUTERS_REQUEST',
    FETCH_COMPUTERS_SUCCESS = 'FETCH_COMPUTERS_SUCCESS',
    FETCH_COMPUTERS_FAILURE = 'FETCH_COMPUTERS_FAILURE',

    CREATE_COMPUTER_REQUEST = 'CREATE_COMPUTER_REQUEST',
    CREATE_COMPUTER_SUCCESS = 'CREATE_COMPUTER_SUCCESS',
    CREATE_COMPUTER_FAILURE = 'CREATE_COMPUTER_FAILURE',

    UPDATE_COMPUTER_REQUEST = 'UPDATE_COMPUTER_REQUEST',
    UPDATE_COMPUTER_SUCCESS = 'UPDATE_COMPUTER_SUCCESS',
    UPDATE_COMPUTER_FAILURE = 'UPDATE_COMPUTER_FAILURE',

    DELETE_COMPUTER_REQUEST = 'DELETE_COMPUTER_REQUEST',
    DELETE_COMPUTER_SUCCESS = 'DELETE_COMPUTER_SUCCESS',
    DELETE_COMPUTER_FAILURE = 'DELETE_COMPUTER_FAILURE',

    GET_SINGLE_COMPUTER_REQUEST = 'GET_SINGLE_COMPUTER_REQUEST',
    GET_SINGLE_COMPUTER_SUCCESS = 'GET_SINGLE_COMPUTER_SUCCESS',
    GET_SINGLE_COMPUTER_FAILURE = 'GET_SINGLE_COMPUTER_FAILURE',
}